feat(LobbyController): extract lobby endpoints
Trennt die Room-Lifecycle-Endpoints (/join, /init, /reconnect, /leave)
aus dem bisherigen WebSocketBrokerController in einen eigenen
Controller. Verhaltensneutral; Endpoint-URLs unverändert.
